modified:   telescope.py 	modified:   utils.py

+++ bayescal/utils.py
@@ -15,7 +15,6 @@ import copy
 viewreal = torch.view_as_real
 viewcomp = torch.view_as_complex
 
-
 def cmult(a, b):
     """
     Complex multiplication of two real-valued torch
@@ -658,6 +657,6 @@ def dynamic_pixelization(base_nside, max_nside, sigma=None, bsky=None, target_ns
     # turn nsides into mhealpy HealpixMap object
     ipix = [healpy.ang2pix(ns, th, ph, nest=True) for ns, th, ph in zip(nsides, theta, phi)]
     uniq = [4 * ns**2 + ip for ns, ip in zip(nsides, ipix)]
-    nsides = mhealpy.HealpixMap(nsides, uniq=uniq, scheme='nest', dtype=np.float32)
+    nsides = mhealpy.HealpixMap(nsides, uniq=uniq, scheme='nest', dtype=np.int16)
 
     return theta, phi, nsides, total_nsides
